+++ examples/room_shoebox_3d.py
@@ -12,12 +12,9 @@ import numpy as np
 
 import pyroomacoustics as pra
 
-# room dimension
-room_dim = [5, 4, 6]
-
 # Create the shoebox
 shoebox = pra.ShoeBox(
-    room_dim,
+    [5, 4, 6],  # room dimension
     absorption=0.2,
     fs=16000,
     max_order=15,
@@ -25,9 +22,7 @@ shoebox = pra.ShoeBox(
 
 # source and mic locations
 shoebox.add_source([2, 3.1, 2])
-shoebox.add_microphone_array(
-        pra.MicrophoneArray(np.array([[2, 1.5, 2]]).T, shoebox.fs)
-        )
+shoebox.add_microphone([2, 1.5, 2])
 
 # run ism
 shoebox.compute_rir()