+++ pyroomacoustics/__init__.py
@@ -123,7 +123,6 @@ from .room import *
 from .beamforming import *
 from .soundsource import *
 from .parameters import *
-from .stft import *
 from .utilities import *
 from .windows import *
 from .sync import *